requirements.txt
streamlit
pillow
tensorflow
openai
requests
fpdf
